+++ backend/tests/test_gb18030.c
@@ -155,7 +155,7 @@ static void test_gb18030_utf8tomb(int index) {
         char *data;
         int length;
         int ret;
-        size_t ret_length;
+        int ret_length;
         unsigned int expected_gbdata[30];
         char *comment;
     };
@@ -190,12 +190,12 @@ static void test_gb18030_utf8tomb(int index) {
         if (index != -1 && i != index) continue;
 
         int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
-        size_t ret_length = length;
+        int ret_length = length;
 
         ret = gb18030_utf8tomb(&symbol, (unsigned char *) data[i].data, &ret_length, gbdata);
         assert_equal(ret, data[i].ret, "i:%d ret %d != %d (%s)\n", i, ret, data[i].ret, symbol.errtxt);
         if (ret == 0) {
-            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %zu != %zu\n", i, ret_length, data[i].ret_length);
+            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %d != %d\n", i, ret_length, data[i].ret_length);
             for (int j = 0; j < (int) ret_length; j++) {
                 assert_equal(gbdata[j], data[i].expected_gbdata[j], "i:%d gbdata[%d] 0x%04X != 0x%04X\n", i, j, gbdata[j], data[i].expected_gbdata[j]);
             }
@@ -216,7 +216,7 @@ static void test_gb18030_utf8tosb(int index) {
         char *data;
         int length;
         int ret;
-        size_t ret_length;
+        int ret_length;
         unsigned int expected_gbdata[30];
         char *comment;
     };
@@ -265,12 +265,12 @@ static void test_gb18030_utf8tosb(int index) {
         if (index != -1 && i != index) continue;
 
         int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
-        size_t ret_length = length;
+        int ret_length = length;
 
         ret = gb18030_utf8tosb(data[i].eci, (unsigned char *) data[i].data, &ret_length, gbdata, data[i].full_multibyte);
         assert_equal(ret, data[i].ret, "i:%d ret %d != %d\n", i, ret, data[i].ret);
         if (ret == 0) {
-            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %zu != %zu\n", i, ret_length, data[i].ret_length);
+            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %d != %d\n", i, ret_length, data[i].ret_length);
             for (int j = 0; j < (int) ret_length; j++) {
                 assert_equal(gbdata[j], data[i].expected_gbdata[j], "i:%d gbdata[%d] %04X != %04X\n", i, j, gbdata[j], data[i].expected_gbdata[j]);
             }
@@ -289,7 +289,7 @@ static void test_gb18030_cpy(int index) {
         char *data;
         int length;
         int ret;
-        size_t ret_length;
+        int ret_length;
         unsigned int expected_gbdata[30];
         char *comment;
     };
@@ -316,10 +316,10 @@ static void test_gb18030_cpy(int index) {
         if (index != -1 && i != index) continue;
 
         int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
-        size_t ret_length = length;
+        int ret_length = length;
 
         gb18030_cpy((unsigned char *) data[i].data, &ret_length, gbdata, data[i].full_multibyte);
-        assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %zu != %zu\n", i, ret_length, data[i].ret_length);
+        assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %d != %d\n", i, ret_length, data[i].ret_length);
         for (int j = 0; j < (int) ret_length; j++) {
             assert_equal(gbdata[j], data[i].expected_gbdata[j], "i:%d gbdata[%d] %04X != %04X\n", i, j, gbdata[j], data[i].expected_gbdata[j]);
         }
