+++ backend/tests/test_gb2312.c
@@ -95,7 +95,7 @@ static void test_gb2312_utf8tomb(int index) {
         char *data;
         int length;
         int ret;
-        size_t ret_length;
+        int ret_length;
         unsigned int expected_gbdata[20];
         char *comment;
     };
@@ -130,12 +130,12 @@ static void test_gb2312_utf8tomb(int index) {
         if (index != -1 && i != index) continue;
 
         int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
-        size_t ret_length = length;
+        int ret_length = length;
 
         ret = gb2312_utf8tomb(&symbol, (unsigned char *) data[i].data, &ret_length, gbdata);
         assert_equal(ret, data[i].ret, "i:%d ret %d != %d (%s)\n", i, ret, data[i].ret, symbol.errtxt);
         if (ret == 0) {
-            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %zu != %zu\n", i, ret_length, data[i].ret_length);
+            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %d != %d\n", i, ret_length, data[i].ret_length);
             for (int j = 0; j < (int) ret_length; j++) {
                 assert_equal(gbdata[j], data[i].expected_gbdata[j], "i:%d gbdata[%d] %04X != %04X\n", i, j, gbdata[j], data[i].expected_gbdata[j]);
             }
@@ -156,7 +156,7 @@ static void test_gb2312_utf8tosb(int index) {
         char *data;
         int length;
         int ret;
-        size_t ret_length;
+        int ret_length;
         unsigned int expected_gbdata[20];
         char *comment;
     };
@@ -203,12 +203,12 @@ static void test_gb2312_utf8tosb(int index) {
         if (index != -1 && i != index) continue;
 
         int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
-        size_t ret_length = length;
+        int ret_length = length;
 
         ret = gb2312_utf8tosb(data[i].eci, (unsigned char *) data[i].data, &ret_length, gbdata, data[i].full_multibyte);
         assert_equal(ret, data[i].ret, "i:%d ret %d != %d\n", i, ret, data[i].ret);
         if (ret == 0) {
-            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %zu != %zu\n", i, ret_length, data[i].ret_length);
+            assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %d != %d\n", i, ret_length, data[i].ret_length);
             for (int j = 0; j < (int) ret_length; j++) {
                 assert_equal(gbdata[j], data[i].expected_gbdata[j], "i:%d gbdata[%d] %04X != %04X\n", i, j, gbdata[j], data[i].expected_gbdata[j]);
             }
@@ -227,7 +227,7 @@ static void test_gb2312_cpy(int index) {
         char *data;
         int length;
         int ret;
-        size_t ret_length;
+        int ret_length;
         unsigned int expected_gbdata[20];
         char *comment;
     };
@@ -254,10 +254,10 @@ static void test_gb2312_cpy(int index) {
         if (index != -1 && i != index) continue;
 
         int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
-        size_t ret_length = length;
+        int ret_length = length;
 
         gb2312_cpy((unsigned char *) data[i].data, &ret_length, gbdata, data[i].full_multibyte);
-        assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %zu != %zu\n", i, ret_length, data[i].ret_length);
+        assert_equal(ret_length, data[i].ret_length, "i:%d ret_length %d != %d\n", i, ret_length, data[i].ret_length);
         for (int j = 0; j < (int) ret_length; j++) {
             assert_equal(gbdata[j], data[i].expected_gbdata[j], "i:%d gbdata[%d] %04X != %04X\n", i, j, gbdata[j], data[i].expected_gbdata[j]);
         }
